+++ src/app/demo/pages/introduction/introduction-demo.component.ts
@@ -13,7 +13,7 @@ import { CodeSnippetComponent } from '../../shared/code-snippet.component';
 })
 export class IntroductionDemoComponent {
   npmUrl = 'https://www.npmjs.com/package/ps-helix';
-  npmVersion = '7.0.0';
+  npmVersion = '7.0.1';
   keyFeatures = [
     {
       icon: 'lightning',